Progress on Dimension
This is all I could do. Had very little time tonight to work on the
mod, and most of it was spent procrastinating and playing Minecraft.

Some stuff is commented out, 'cause work on actually getting that stuff
to compile is being made.

Also, the file isn't really called "mod10k.java" anymore. It wasn't
since the commit before this one. It's called "Mod10K.java" like the
contents of the file would suggest. It just seems that GitHub doesn't
seem to recognize changes in the capitalization of file names(?)