Fixed portal crash, removed item teleporter, and re-implemented portal
Item teleporter was always temporary, and I'm a happy man to see it go.
I also re-added the portal, which is made basically the same way as the
Nether portal, only you use the portal trigger instead of fire.

Now, there's a bug (not a crash! :D ) where when you enter the portal,
it does take you to 10K, but it's a void dimension with no terrain
whatsoever, and immediatly teleports you again to the Nether. Odd, huh?